+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 ## 4.33.0 (Unreleased)
+
+FEATURES:
+
+* **New Resource:** `aws_identitystore_group` ([#26674](https://github.com/hashicorp/terraform-provider-aws/issues/26674))
+* **New Resource:** `aws_identitystore_user` ([#26948](https://github.com/hashicorp/terraform-provider-aws/issues/26948))
+* **New Resource:** `aws_inspector2_organization_configuration` ([#27000](https://github.com/hashicorp/terraform-provider-aws/issues/27000))
+
+BUG FIXES:
+
+* resource/aws_ssm_parameter: Allow parameter overwrite on create ([#26785](https://github.com/hashicorp/terraform-provider-aws/issues/26785))
+
 ## 4.32.0 (September 23, 2022)
 
 ENHANCEMENTS: